D3ASIM-3553-market-agent: Rename InterAreaAgent to MarketAgent
