+++ src/gsy_e/models/area/event_dispatcher.py
@@ -57,16 +57,16 @@ class AreaDispatcher:
     dicts with interarea agents for each market type.
     """
     def __init__(self, area: "Area"):
-        self._inter_area_agents: Dict[DateTime, OneSidedAgent] = {}
+        self._market_agents: Dict[DateTime, OneSidedAgent] = {}
         self._balancing_agents: Dict[DateTime, BalancingAgent] = {}
         self._settlement_agents: Dict[DateTime, SettlementAgent] = {}
         self._future_agent: Optional[FutureAgent] = None
         self.area = area
 
     @property
-    def interarea_agents(self) -> Dict[DateTime, OneSidedAgent]:
+    def market_agents(self) -> Dict[DateTime, OneSidedAgent]:
         """Return spot market inter area agents."""
-        return self._inter_area_agents
+        return self._market_agents
 
     @property
     def balancing_agents(self) -> Dict[DateTime, BalancingAgent]:
@@ -257,7 +257,7 @@ class AreaDispatcher:
             dispatcher_object, market_type: AvailableMarketTypes
     ) -> Dict[DateTime, Union[OneSidedAgent, BalancingAgent, SettlementAgent]]:
         if market_type == AvailableMarketTypes.SPOT:
-            return dispatcher_object.interarea_agents
+            return dispatcher_object.market_agents
         if market_type == AvailableMarketTypes.BALANCING:
             return dispatcher_object.balancing_agents
         if market_type == AvailableMarketTypes.SETTLEMENT:
@@ -304,10 +304,10 @@ class AreaDispatcher:
         if not self._should_agent_be_created:
             return
 
-        interarea_agents = self._get_agents_for_market_type(self, market_type)
+        market_agents = self._get_agents_for_market_type(self, market_type)
         parent_markets = self.area.parent.get_market_instances_from_class_type(
             market_type)
-        if market.time_slot in interarea_agents or market.time_slot not in parent_markets:
+        if market.time_slot in market_agents or market.time_slot not in parent_markets:
             return
 
         iaa = self._create_agent_object(
@@ -318,11 +318,11 @@ class AreaDispatcher:
         )
 
         # Attach agent to own IAA dict
-        interarea_agents[market.time_slot] = iaa
+        market_agents[market.time_slot] = iaa
 
     def event_market_cycle(self) -> None:
         """Called every market cycle. Recycles old area agents."""
-        self._delete_past_agents(self._inter_area_agents)
+        self._delete_past_agents(self._market_agents)
         self._delete_past_agents(self._balancing_agents)
         self._delete_past_agents(self._settlement_agents)
