+++ tests/test_balancing_agent.py
@@ -29,7 +29,7 @@ from gsy_framework.data_classes import BalancingOffer, BalancingTrade, Offer, Tr
 from gsy_e.constants import TIME_ZONE
 from gsy_e.gsy_e_core.exceptions import InvalidBalancingTradeException
 from gsy_e.models.strategy.area_agents.balancing_agent import BalancingAgent
-from tests.test_inter_area_agent import FakeMarket
+from tests.test_market_agent import FakeMarket
 
 
 class FakeArea:
