+++ tests/test_market_agent.py
